requirements.txt
aiogram==3.7.0
anthropic>=0.39.0
aiosqlite>=0.19.0
Pillow>=10.0.0
pyzbar>=0.1.9
python-dotenv>=1.0.0
bcrypt>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
flask>=3.0.0
